implement conversion from IP to byte arrays
This is very close to the included `.octets()` method but it enables
declaring trait bounds for `Into<[u8; N]>` on `IpvNAddr`. It is also
symmetric with the existing `From<[u8; N]>` counterparts which are
already there.

Converting IP addresses to/from `[u8]` is a natural thing in the
networking world since this is how they are written to and read from
packets and how humans parse them. It also happens to be the same as the
internal implementation of the IpvNAddr structs which should make it an
efficient operation. But, that is inconsequential from an API point of
view, it is just an aside.

As an example of real-world use, I'm currently starting code to
implement IP sets and tables which will store IP address keys in a
generic way, only requiring that they be convertible to and from the
fixed-sized array representation. They'll be stored efficiently in
special-purpose, [u8]-based, trie data structures. I have looked through
current Rust implementations of IP-trie data structures and found that
they don't quite meet my needs.