copy rustlib sources instead of using symbolic links
There are certain cases where we need to add things to the rustlib source
paths in CI builders. When the rustlib sources are symbolically linked to
`/checkout`, we can't make any modifications because `/checkout` is mounted
as read-only in containers.
This change updates the bootstrap to copy the rustlib directories instead of
linking them, so we can put/link things into rustlib directories when needed.

Signed-off-by: onur-ozkan <work@onurozkan.dev>