Change to the NetBSD archive URL rather than the CDN
The CDN has been down for a few hours. Switch to an alternative for the
time being so we can unblock CI.

It appears that the CDN is quite a bit faster, so we will likely want to
switch back when available.

(cherry picked from commit 42ac903f4fe4e9894f8c6e97538696bdbae848dc)