linker: when emitting static executables, explicitly hint static
This is important as rustc is not always in charge of what it is
linking (due to some targets not necessarily passing -nodefaultlibs)
and it assumes that resetting to dynamic is always the default.

This would result in executables that have no interpreter but
are still dynamically linked (despite static being requested).

By explicitly hinting static again before passing the -static
or -static-pie, we tell the linker to always link whatever
implicit libs statically as expected.