Normalize revisions used for filecheck directives
Typically, filecheck prefixes will be uppercase (always true) and start
with `CHECK-` (almost always true). Currently we allow using revision
names as filecheck directives, but they are passed directly.  That means
that our directives are exactly what the revision name is, in the same
case, so they only look like filecheck directives if the revision name
is uppercase (usually they are lowercase).

Update this so that we always uppercase revision names and prefix them
with `CHECK-` when used as directives. This is better for consistency,
makes it easier to identify directives in the tests, and has the nice
side effect that some editors will make the directive stand out by
highlighting it (which currently happens for most directives, just not
those from revisions).