Revert "Make `thir::Pat` not implement `fmt::Display` directly"
This reverts commit ae0ec731a86ac6a2f6d8d08c4996bb8a438afdc2.

The original change in #128304 was intended to be a step towards being able to
print `thir::Pat` even after switching to `PatId`.

But because the only patterns that need to be printed are the synthetic ones
created by pattern analysis (for diagnostic purposes only), it makes more sense
to completely separate the printable patterns from the real THIR patterns.