Enable `std/panic-unwind` when checking `std`
`std/panic-unwind` gets enabled when building the sysroot, but not when
checking `std`. This was not a problem with the v1 resolver because it
would unify features and just always enable `panic-unwind`.

With the v2 resolver, however, this causes `std` to get built twice with
different sets of features. This then causes an "multiple candidates for
`rmeta` dependency `std` found" error.

Always enable the `panic-unwind` feature to avoid this conflict.